Updates clean/blacklisting functionality

+++ bin/clean.sh
@@ -1,9 +1,23 @@
 #!/usr/bin/env bash
 
 BLACKLIST='blacklist.txt'
+BLACKLIST_GLOBAL='../blacklist.txt'
 
-readarray -t idList < "$BLACKLIST"
+if [ -f "$BLACKLIST" ];then
+    readarray -t idList < "$BLACKLIST"
 
-for id in "${idList[@]}";do
-    find . -type d -iname "${id}*"  -print0 | xargs -0 rm -rfv
-done
+    for id in "${idList[@]}";do
+        find . -type d -iname "${id}*"  -print0 | xargs -0 rm -rfv
+    done
+fi
+
+if [ -f "$BLACKLIST_GLOBAL" ];then
+    readarray -t idListGlobal < "$BLACKLIST_GLOBAL"
+
+    for id in "${idListGlobal[@]}";do
+        if [ -z "$id" ];then
+            continue
+        fi
+        find . -type d -iname "${id}*"  -print0 | xargs -0 rm -rfv
+    done
+fi
